Crate tidb_query_expr

This crate implements a simple SQL query engine to work with TiDB pushed down executors.

The query engine is able to scan and understand rows stored by TiDB, run against a series of executors and then return the execution result. The query engine is provided via TiKV Coprocessor interface. However standalone UDF functions are also exported and can be used standalone.

Re-exports

pub use self::types::*;

Modules

impl_arithmetic

impl_cast

impl_compare

impl_compare_in

impl_control

impl_encryption

impl_json

impl_like

impl_math

impl_miscellaneous

impl_op

impl_other

impl_string

impl_time

types

Functions

compare_mapper

divide_mapper

map_compare_in_string_sig

map_expr_node_to_rpn_func

map_int_sig

map_like_sig

map_rhs_int_sig

map_string_compare_sig

map_unary_minus_int_func

minus_mapper

mod_mapper

multiply_mapper

plus_mapper

truncate_int_mapper

truncate_real_mapper