Module tidb_query_expr::types

Re-exports

pub use self::function::RpnFnCallExtra;

pub use self::function::RpnFnMeta;

Modules

expr

expr_builder

expr_eval

function
People implementing RPN functions with fixed argument type and count don't necessarily need to understand how Evaluator and RpnDef work. There's a procedural macro called rpn_fn defined in tidb_query_codegen to help you create RPN functions. For example:

Structs

RpnExpression
An expression in Reverse Polish notation, which is simply a list of RPN expression nodes.

RpnExpressionBuilder
Helper to build an RpnExpression.

Enums

RpnExpressionNode
A type for each node in the RPN expression list.

RpnStackNode
A type for each node in the RPN evaluation stack. It can be one of a scalar value node or a vector value node. The vector value node can be either an owned vector value or a reference.

Constants

BATCH_MAX_SIZE