Trait tidb_query_expr::types::function::Evaluator

pub trait Evaluator<'a> {
    fn eval(
        self, 
        def: impl ArgDef, 
        ctx: &mut EvalContext, 
        output_rows: usize, 
        args: &'a [RpnStackNode<'a>], 
        extra: &mut RpnFnCallExtra<'_>, 
        metadata: &(dyn Any + Send)
    ) -> Result<VectorValue>;
}

A generic evaluator of an RPN function.

For every RPN function, the evaluator should be created first. Then, call its eval method with the input to get the result vector.

There are two kinds of evaluators in general:

ArgConstructor: It's a provided Evaluator. It is used in the rpn_fn attribute macro to generate the ArgDef. The def parameter of its eval method is the already constructed ArgDef. If it is the outmost evaluator, def should be Null.

Custom evaluators which do the actual execution of the RPN function. The def parameter of its eval method is the constructed ArgDef. Implementors can then extract values from the arguments, execute the RPN function and fill the result vector.

Required methods

fn eval(
self,
def: impl ArgDef,
ctx: &mut EvalContext,
output_rows: usize,
args: &'a [RpnStackNode<'a>],
extra: &mut RpnFnCallExtra<'_>,
metadata: &(dyn Any + Send)
) -> Result<VectorValue>

Loading content...